Start moving things to a printing module
